+++ packages/create-cloudflare/e2e/tests/frameworks/test-config.ts
@@ -148,6 +148,9 @@ function getFrameworkTestConfig(pm: string): NamedFrameworkTestConfig[] {
 		},
 		{
 			name: "analog",
+			// @analogjs/vite-plugin-angular is incompatible with @angular/build 22.2.0.
+			// See https://github.com/analogjs/analog/issues/2574.
+			quarantine: true,
 			testCommitMessage: true,
 			timeout: LONG_TIMEOUT,
 			unsupportedOSs: ["win32"],
@@ -1102,10 +1105,6 @@ function getExperimentalFrameworkTestConfig(
 		},
 		{
 			name: "redwood",
-			// quarantined because redwood v1.7.4 has failed to be released but that is currently
-			// the version that create-rwsdk tried to use
-			// TODO: unquarantine once the upstream redwood issue is resolved
-			quarantine: true,
 			testCommitMessage: true,
 			timeout: LONG_TIMEOUT,
 			unsupportedOSs: ["win32"],
@@ -1126,6 +1125,9 @@ function getExperimentalFrameworkTestConfig(
 		},
 		{
 			name: "analog",
+			// @analogjs/vite-plugin-angular is incompatible with @angular/build 22.2.0.
+			// See https://github.com/analogjs/analog/issues/2574.
+			quarantine: true,
 			testCommitMessage: true,
 			timeout: LONG_TIMEOUT,
 			unsupportedOSs: ["win32"],